add crud operations for Items

+++ src/components/List.js
@@ -1,12 +1,17 @@
 import { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import API_URL from '../data/api';
-import { Accordion, Button } from 'react-bootstrap';
+import { Accordion, Button, Form } from 'react-bootstrap';
+import Item from './Item';
 
 const List = () => {
   const { listId } = useParams();
   const [list, setList] = useState({});
   const [items, setItems] = useState([]);
+  const [isAddingNewItem, setIsAddingNewItem] = useState(false);
+  const [title, setTitle] = useState('');
+  const [text, setText] = useState('');
+  const navigate = useNavigate();
 
   useEffect(() => {
     const getList = async () => {
@@ -23,7 +28,51 @@ const List = () => {
     getList();
   }, [listId]);
 
-  const handleDelete = async (itemId) => {
+  const handleSubmit = (e) => {
+    e.preventDefault();
+
+    if (title === '') {
+      alert('The new item must have a title');
+      return;
+    }
+
+    createItem({ title, text });
+    resetForm();
+  };
+
+  const resetForm = () => {
+    setIsAddingNewItem(false);
+    setTitle('');
+    setText('');
+  };
+
+  const createItem = async (newItem) => {
+    const response = await fetch(`${API_URL}/${listId}/items`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(newItem),
+    });
+    const data = await response.json();
+    setItems(items.concat(data));
+  };
+
+  const updateItem = async (itemToUpdate) => {
+    setItems(
+      items.map((item) =>
+        item.id === itemToUpdate.id ? { ...item, ...itemToUpdate } : item
+      )
+    );
+    await fetch(`${API_URL}/${listId}/items/${itemToUpdate.id}`, {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(itemToUpdate),
+    });
+  };
+
+  const deleteItem = async (itemId) => {
+    if (!window.confirm('Are you sure you want to delete this item?')) {
+      return;
+    }
     setItems(items.filter((item) => item.id !== itemId));
     await fetch(`${API_URL}/${listId}/items/${itemId}`, {
       method: 'DELETE',
@@ -32,23 +81,80 @@ const List = () => {
 
   return (
     <>
-      <h1 className="my-3">{list.name}</h1>
+      <h1 className="my-3">
+        {list.name}
+        <Button
+          variant="primary"
+          className="mx-3"
+          onClick={() => navigate('./edit')}
+        >
+          Edit List
+        </Button>
+        <Button
+          variant="danger"
+          onClick={() => window.confirm('Delete list and all its items?')}
+        >
+          Delete List
+        </Button>
+      </h1>
+
+      {isAddingNewItem ? (
+        <>
+          <Form onSubmit={handleSubmit}>
+            <h4>Add New Item</h4>
+            <Form.Group className="mb-3" controlId="title">
+              <Form.Label>Title</Form.Label>
+              <Form.Control
+                type="text"
+                value={title}
+                onChange={(e) => setTitle(e.target.value)}
+                placeholder="Enter the item title"
+              />
+            </Form.Group>
+            <Form.Group className="mb-3" controlId="text">
+              <Form.Label>Text</Form.Label>
+              <Form.Control
+                as="textarea"
+                value={text}
+                onChange={(e) => setText(e.target.value)}
+                placeholder="Enter the item text"
+              />
+            </Form.Group>
+            <Button
+              variant="secondary"
+              type="button"
+              className="me-3 mb-3"
+              onClick={() => {
+                setIsAddingNewItem(false);
+                setTitle('');
+                setText('');
+              }}
+            >
+              Cancel
+            </Button>
+            <Button variant="primary" type="submit" className="mb-3">
+              Submit
+            </Button>
+          </Form>
+        </>
+      ) : (
+        <Button
+          variant="primary"
+          className="my-3"
+          onClick={() => setIsAddingNewItem(true)}
+        >
+          New Item
+        </Button>
+      )}
 
       <Accordion>
-        {items.map((item, index) => (
-          <Accordion.Item eventKey={index} key={item.id}>
-            <Accordion.Header>{item.title}</Accordion.Header>
-            <Accordion.Body>
-              <p>{item.text}</p>
-              <Button
-                onClick={() => {
-                  handleDelete(item.id);
-                }}
-              >
-                Delete me
-              </Button>
-            </Accordion.Body>
-          </Accordion.Item>
+        {items.map((item) => (
+          <Item
+            key={item.id}
+            item={item}
+            deleteItem={deleteItem}
+            updateItem={updateItem}
+          />
         ))}
       </Accordion>
     </>
